requirements.txt
aiogram==3.13.1
python-docx==1.1.0
openai==1.42.0
python-dotenv==1.0.1
httpx==0.27.2
aiohttp
requests